+++ safe/gis/vector/summary_1_aggregate_hazard.py
@@ -26,7 +26,7 @@ from safe.definitions.utilities import definition
 from safe.gis.sanity_check import check_layer
 from safe.gis.vector.summary_tools import (
     check_inputs, create_absolute_values_structure, add_fields)
-from safe.post_processors import post_processor_affected_function
+from safe.processors import post_processor_affected_function
 from safe.utilities.gis import qgis_version
 from safe.utilities.i18n import tr
 from safe.utilities.pivot_table import FlatTable
