+++ safe/gis/vector/summary_3_analysis.py
@@ -27,7 +27,7 @@ from safe.definitions.processing_steps import (
 from safe.gis.sanity_check import check_layer
 from safe.gis.vector.summary_tools import (
     check_inputs, create_absolute_values_structure, add_fields)
-from safe.post_processors import post_processor_affected_function
+from safe.processors import post_processor_affected_function
 from safe.utilities.gis import qgis_version
 from safe.utilities.pivot_table import FlatTable
 from safe.utilities.profiling import profile
