+++ safe/impact_function/postprocessors.py
@@ -9,7 +9,7 @@ from qgis.core import QgsFeatureRequest
 from safe.definitions.minimum_needs import minimum_needs_parameter
 from safe.gis.vector.tools import (
     create_field_from_definition, SizeCalculator)
-from safe.post_processors import (
+from safe.processors import (
     field_input_type,
     keyword_input_type,
     keyword_value_expected,
